ref(ui): Move integration external components to integrations
Just moving these out of the components directory since they aren't used outside of integrations
